- Problem of slave's transparency solved : we were creating each time a new QStretchTilerItem for each call of 'scene/sync slave master'. The finding of the master of the already existing masters was not working, now it is. - We switch to float for the 'relative2Scene' methods : the float-double conversions were creating graphical errors on the export.
Former-commit-id: fd68a50472779e30318374a2a47e568196307d2b

+++ src/view/VDummyObjectView.h
@@ -55,24 +55,24 @@ class VDummyObjectView : public VObjectView
 		virtual GraphicEffect getEffect () const			{ return GraphicEffect(); }
     
 		/// \brief Maps the IObject [-1,1] y coordinate to the referenceRect().
-		virtual double relative2SceneY(float y, QGraphicsItem * item = 0 ) const						{ return y; }
+		virtual float relative2SceneY(float y, QGraphicsItem * item = 0 ) const						{ return y; }
 		/// \brief Maps the IObject [-1,1] x coordinate to the referenceRect().
-		virtual double relative2SceneX(float x, QGraphicsItem * item = 0 ) const						{ return x; }
+		virtual float relative2SceneX(float x, QGraphicsItem * item = 0 ) const						{ return x; }
 		/// \brief Maps the IObject [0,2] width value to the corresponding referenceRect() value.
-		virtual double relative2SceneWidth(float width, QGraphicsItem * item = 0 ) const				{ return width; }
+		virtual float relative2SceneWidth(float width, QGraphicsItem * item = 0 ) const				{ return width; }
 		/// \brief Maps the IObject [0,2] height value to the corresponding referenceRect() value.
-		virtual double relative2SceneHeight(float height, QGraphicsItem * item = 0 ) const				{ return height; }
+		virtual float relative2SceneHeight(float height, QGraphicsItem * item = 0 ) const				{ return height; }
 		/// \brief Maps a rect expressed in [-1,1] scene coordinate to a QRectF expressed in referenceRect() coordinates.
 //		virtual QRectF relative2SceneRect( const TFloatRect& rect) const	= 0;
 
 		/// \brief Maps the referenceRect() width value to the corresponding [0,2] value.
-		virtual double scene2RelativeWidth(float width, QGraphicsItem * item = 0 ) const				{ return width; }
+		virtual float scene2RelativeWidth(float width, QGraphicsItem * item = 0 ) const				{ return width; }
 		/// \brief Maps the referenceRect() height value to the corresponding [0,2] value.
-		virtual double scene2RelativeHeight(float height, QGraphicsItem * item = 0 ) const				{ return height; }
+		virtual float scene2RelativeHeight(float height, QGraphicsItem * item = 0 ) const				{ return height; }
 		/// \brief Maps the referenceRect() x value to the corresponding [-1,1] value.
-		virtual double scene2RelativeX(float x, QGraphicsItem * item = 0 ) const						{ return x; }
+		virtual float scene2RelativeX(float x, QGraphicsItem * item = 0 ) const						{ return x; }
 		/// \brief Maps the referenceRect() y value to the corresponding [-1,1] value.
-		virtual double scene2RelativeY(float y, QGraphicsItem * item = 0 ) const						{ return y; }
+		virtual float scene2RelativeY(float y, QGraphicsItem * item = 0 ) const						{ return y; }
 };
 
 /*!@} */
